src/app_predictor/requirements.txt
opencv-python==4.6.0.66
redis==4.3.4
requests==2.28.1
python-telegram-bot==13.14